+++ tests/test_data/@ClassFolderUsingBuiltins/ClassFolderUsingBuiltins.m
@@ -1,0 +1,8 @@
+classdef ClassFolderUsingBuiltins
+    % The `angle` is a builtin name in MATLAB
+
+    methods
+        angle = GetAngle(obj);
+        length = GetLength(obj);
+    end
+end