+++ tests/test_data/@NestedClassFolder/private/private_function.m
@@ -1,0 +1,3 @@
+function y = private_function(x)
+% A private function
+y = x;